Add longtasks buffering support (#8)
Based on the discussion from the latest WebPerf call, we've decided to allow the buffered flag to work with longtasks. We've set a maxBufferSize of 200 tentatively. It's likely going to be enough but we can consider increasing it in the future if people are hitting the 200 limit when getting entries from the PeformanceObserver.

With this change, we only need to change the longtasks spec to remove a Note saying